+++ src/main/java/racingcar/service/GameService.java
@@ -4,6 +4,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import racingcar.domain.Car;
+import racingcar.domain.Cars;
+import racingcar.domain.Lap;
+import racingcar.domain.NumberGenerator;
+import racingcar.domain.RandomNumberGenerator;
+import racingcar.domain.WinnerMaker;
+import racingcar.dto.request.GameRequestDto;
 import racingcar.dto.response.GameResponseDto;
 import racingcar.dto.request.GameResultDto;
 import racingcar.dto.request.GameSaveDto;
@@ -19,6 +25,9 @@ import java.util.List;
 @Service
 public class GameService {
 
+    private static final int MINIMUM_RANDOM_NUMBER = 0;
+    private static final int MAXIMUM_RANDOM_NUMBER = 9;
+
     private final GameRepository gameRepository;
     private final PlayerResultRepository playerResultRepository;
 
@@ -29,10 +38,25 @@ public class GameService {
     }
 
     @Transactional
-    public GameResponseDto playGame(GameResultDto gameResult) {
+    public GameResponseDto playGame(final GameRequestDto request) {
+        final Cars cars = new Cars(request.getNames());
+        final Lap lap = new Lap(request.getCount());
+        race(cars, lap, new RandomNumberGenerator(MINIMUM_RANDOM_NUMBER, MAXIMUM_RANDOM_NUMBER));
+
+        final WinnerMaker winnerMaker = new WinnerMaker();
+        List<String> winners = winnerMaker.getWinnerCarsName(cars.getLatestResult());
+
+        final GameResultDto gameResult = GameResultDto.of(winners, request.getCount(), cars.getLatestResult());
         final Game game = createGame(gameResult);
         final List<PlayerResult> playerResults = createPlayerResults(gameResult, game);
         return GameResponseDto.of(game, playerResults);
+    }
+
+    private void race(Cars cars, Lap lap, NumberGenerator numberGenerator) {
+        while (!lap.isFinish()) {
+            cars.moveCars(numberGenerator);
+            lap.reduce();
+        }
     }
 
     private Game createGame(final GameResultDto gameResult) {
